fix: comprehensive security audit remediation (34 issues)
Critical fixes:
- C-1: Replace all int64*1e18 overflow with sdkmath.Int/big.Int
- C-2: Remove plaintext question from on-chain state, use hash-only lookup
- C-3: Add reveal time window (must be after commit deadline, before eval)
- C-5: Remove hardcoded SECRET_KEY_BASE, require env var

High-priority fixes:
- H-1: Collect-then-mutate pattern for IterateAgents (abci, reputation)
- H-2: Detect cheaters by normalized reveal content, not just commit hash
- H-3: Epoch boundary tracking via state (survives param changes)
- H-4: IncrementContractCalls now called from EVM hooks
- H-5: Activity from challenge submission no longer counted
- H-6: Full Params.Validate() for all fields
- H-7: Precompile gas metering improvements
- H-8: Remove personal/debug from JSON-RPC API
- H-9: Deprecate --private-key CLI flag in daemon
- H-13: AI reward fallback goes to pool, not proposer
- H-14: Daily registration limit (3 per address per ~24h)

Medium fixes:
- M-1: Gas burn percentage documentation alignment
- M-2: BytesToUint64 length check prevents panic
- M-3: AIBonus bounded to [-5, +30] range
- M-4: Epoch data cleanup on deregister
- M-5: BurnedAtRegister snapshot in Agent struct
- M-6: Epoch state tracking resilient to param changes
- M-7/M-8: Docker security hardening (localhost ports, pinned images)
- M-11/M-12: Daemon exponential backoff + receipt error propagation
- M-13: Python SDK int instead of float for stake amounts

Additional improvements:
- New error codes: ErrRevealTooEarly, ErrRevealWindowClosed, etc.
- Suspended agents cannot call UpdateAgent
- New param validation tests

Made-with: Cursor

+++ sdk/python/axon/client.py
@@ -103,11 +103,11 @@ class AgentClient:
         }
 
     def register_agent(
-        self, capabilities: str, model: str, stake_axon: float = 100
+        self, capabilities: str, model: str, stake_axon: int = 100
     ) -> str:
         """Register as an AI Agent. Requires >= 100 AXON stake (20 burned)."""
         self._require_account()
-        stake_wei = int(stake_axon * ONE_AXON)
+        stake_wei = int(stake_axon) * ONE_AXON
         tx = self._registry.functions.register(capabilities, model).build_transaction(
             self._tx_params(value=stake_wei)
         )
